requirements.txt
pandas==0.24.2
mysql.connector==2.2.9
psycopg2==2.8.3
matplotlib==2.2..4
dash==0.43.0